added tsi test
