+++ jl_hiop_pridec/.gitignore
@@ -1,3 +1,4 @@
+.DS_Store
 build
 output
 config.sh
