+++ jl_hiop_pridec/julia_src/test.jl
@@ -1,0 +1,6 @@
+
+
+include("test_cases.jl")
+ptr=test_9bus()
+sol=solve_base_case(ptr)
+
