+++ jl_hiop_pridec/julia_src/test_cases.jl
@@ -1,0 +1,31 @@
+
+include("hiop.jl")
+
+function test_cali()
+
+    prob="/usr/workspace/santiago/EXAJUJO/SCACOPF/exajugo/examples/California";
+    rawfile=joinpath(prob,"case.raw"); confile=joinpath(prob,"case.con"); ropfile=joinpath(prob,"case.rop");
+
+    confile=joinpath(prob,"case_Eaton_SD_2.con");
+    return load_ACOPF(rawfile, ropfile, confile)
+
+end
+
+function test_500bus()
+
+    prob="/p/lustre1/santiago/COSMIN/GITHUB/exajugo/examples/500bus/";
+    rawfile=joinpath(prob,"case.raw"); confile=joinpath(prob,"case.con"); ropfile=joinpath(prob,"case.rop");
+    inlfile=joinpath(prob,"case.inl")
+
+    return load_ACOPF_dir(prob)
+
+end
+
+function test_9bus()
+
+    prob="/p/lustre1/santiago/COSMIN/GITHUB/exajugo/examples/9bus/";
+    rawfile=joinpath(prob,"case.raw"); confile=joinpath(prob,"case.con"); ropfile=joinpath(prob,"case.rop");
+
+    return load_ACOPF(rawfile, ropfile, confile)
+
+end
