--- jl_hiop_pridec/test.jl
@@ -1,6 +1,0 @@
-
-
-include("test_cases.jl")
-ptr=test_9bus()
-sol=solve_base_case(ptr)
-
