+++ jl_hiop_pridec/test_tsi/data/ACTIVSg500/case.inl
@@ -1,0 +1,91 @@
+9,1,33.6261,0.8333,0.25,700.0,0.0
+16,1,19.1629,0.8333,0.25,650.0,0.0
+17,1,26.2179,0.8333,0.25,888.84,0.0
+18,1,9.7215,0.8333,0.25,157.6,0.0
+49,1,0.0778,0.8333,0.25,1.3,0.0
+50,1,0.1175,0.8333,0.25,2.6,0.0
+63,1,0.3076,0.8333,0.25,3.2,0.0
+71,1,0.1703,0.1198,0.0,4.0,0.0
+72,1,0.1256,0.1198,0.0,2.0,0.0
+73,1,0.0612,0.1198,0.0,1.4,0.0
+82,1,10.6555,0.1198,0.0,361.9,0.0
+125,1,5.5124,0.1198,0.0,23.29469,0.0
+126,1,5.6149,0.1198,0.0,23.29469,0.0
+127,1,11.5001,0.1198,0.0,162.0,0.0
+128,1,3.2368,0.1198,0.0,162.0,0.0
+144,1,22.1379,0.8333,0.25,602.55,0.0
+145,1,26.3657,0.8333,0.25,602.55,0.0
+167,1,9.8195,0.1198,0.0,193.5,0.0
+168,1,4.8739,0.1198,0.0,193.5,0.0
+169,1,9.1826,0.1198,0.0,387.0,0.0
+197,1,11.9348,0.8333,0.25,293.6,0.0
+198,1,10.377,0.8333,0.25,207.3,0.0
+222,1,0.1092,0.1198,0.0,5.2,0.0
+223,1,0.7505,0.1198,0.0,18.0,0.0
+224,1,15.6709,0.1198,0.0,100.0,0.0
+225,1,24.2032,0.1198,0.0,100.0,0.0
+231,1,0.3002,0.8333,0.25,2.9,0.0
+248,1,9.7566,0.8333,0.25,133.2,0.0
+249,1,3.6077,0.8333,0.25,133.2,0.0
+250,1,19.1015,0.8333,0.25,532.80001,0.0
+254,1,0.2984,0.8333,0.25,4.1,0.0
+255,1,0.3881,0.8333,0.25,3.3,0.0
+256,1,0.1138,0.8333,0.25,2.9,0.0
+257,1,0.6178,0.8333,0.25,5.4,0.0
+258,1,0.1199,0.8333,0.25,1.1,0.0
+264,1,1.2813,0.8333,0.25,39.2,0.0
+265,1,0.9859,0.8333,0.25,10.6,0.0
+275,1,4.5672,0.8333,0.25,101.2,0.0
+301,1,0.0931,0.8333,0.25,2.9,0.0
+302,1,0.1823,0.8333,0.25,2.9,0.0
+305,1,0.0325,0.1198,0.0,1.0,0.0
+306,1,0.0748,0.1198,0.0,2.4,0.0
+319,1,1.7119,0.1198,0.0,42.3,0.0
+350,1,0.49,0.1198,0.0,15.0,0.0
+351,1,0.2529,0.1198,0.0,6.2,0.0
+352,1,0.6646,0.1198,0.0,14.4,0.0
+353,1,5.8106,0.1198,0.0,74.0,0.0
+361,1,0.1347,0.8333,0.25,1.6,0.0
+362,1,0.8036,0.8333,0.25,11.0,0.0
+363,1,0.3551,0.8333,0.25,3.2,0.0
+388,1,0.6594,0.8333,0.25,10.0,0.0
+389,1,12.0499,0.8333,0.25,52.0,0.0
+390,1,12.7658,0.8333,0.25,52.0,0.0
+391,1,10.8699,0.8333,0.25,173.33333,0.0
+392,1,8.9166,0.8333,0.25,116.4,0.0
+393,1,7.1804,0.8333,0.25,116.4,0.0
+394,1,4.6215,0.8333,0.25,116.4,0.0
+395,1,9.3364,0.8333,0.25,116.4,0.0
+396,1,10.3524,0.8333,0.25,116.4,0.0
+409,1,9.1745,0.8333,0.25,100.0,0.0
+410,1,9.1634,0.8333,0.25,41.08,0.0
+411,1,9.6122,0.8333,0.25,41.08,0.0
+412,1,9.0477,0.8333,0.25,41.08,0.0
+413,1,3.8184,0.8333,0.25,41.08,0.0
+414,1,7.2965,0.8333,0.25,111.09999,0.0
+430,1,0.0447,0.1198,0.0,1.0,0.0
+431,1,0.3907,0.1198,0.0,3.3,0.0
+432,1,0.0786,0.1199,0.0,1.2,0.0
+433,1,0.5053,0.1198,0.0,6.7,0.0
+434,1,0.447,0.1198,0.0,4.3,0.0
+435,1,9.4215,0.1198,0.0,28.3275,0.0
+436,1,13.2545,0.1198,0.0,28.3275,0.0
+437,1,4.4384,0.1198,0.0,197.0,0.0
+438,1,6.2703,0.1198,0.0,197.0,0.0
+439,1,8.2324,0.1198,0.0,197.0,0.0
+442,1,8.9834,0.8333,0.25,108.0,0.0
+443,1,12.0059,0.8333,0.25,175.0,0.0
+444,1,0.2255,0.8333,0.25,3.3,0.0
+455,1,5.1504,0.8333,0.25,66.85,0.0
+456,1,13.9085,0.8333,0.25,133.7,0.0
+458,1,6.7146,0.8333,0.25,109.6,0.0
+463,1,0.3546,0.8333,0.25,7.2,0.0
+480,1,2.0839,0.1198,0.0,56.0,0.0
+481,1,1.64,0.1198,0.0,45.0,0.0
+482,1,2.6222,0.1198,0.0,45.0,0.0
+483,1,2.2747,0.1198,0.0,24.0,0.0
+484,1,2.3662,0.1198,0.0,28.0,0.0
+494,1,0.323,0.8333,0.25,5.8,0.0
+497,1,5.1126,0.1198,0.0,60.0,0.0
+498,1,1.3151,0.1198,0.0,18.0,0.0
+0 / END OF DATA
